pyproject.toml
[project]
name = "zerowaste-bootstrap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "pycocotools>=2.0.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "requests>=2.32.5",
    "scipy>=1.17.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "typer[all]>=0.24.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
zerowaste-bootstrap = "zerowaste_bootstrap.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (downloads model weights)",
]
